Almost complete rewrite, use DokuWiki index and parser
This changes the plugin to store the anchors in the metadata in the
index and to execute replacements by generating a pattern for a regular
syntax parts. This fixes all kinds of bugs with incorrect metadata and
section editing. Backlinks, table of contents and useheading should work
without problems. This also makes the replacements a lot more robust,
they won't happen in places anymore where they shouldn't happen but they
should happen in all places where you could also use a normal link syntax.

Furthermore this adds cache handling by letting the parser caches depend
on the anchor index. This ensures that pages are rendered again whenever
the defined anchors are changed.